Update README.md
#### About the project

This project was solely made so people can test the strength of the password they use or would like to use in the future.


#### Goals

In this repo, I will keep it updated with the current source code to the website, I currently am re-working the entire website. 

We currently have a trello of the ideas we plan to either adding in this update or a future update. You can view it [here](https://trello.com/b/nWk64vpZ/hsimp).

#### Feature Request? 

If you have an idea that would benefit this project please go ahead and open up an "Issue" ticket. You can locate that [here](https://github.com/Baker/HSIMP/issues). I will attend to these and add them to our trello, after we get good feedback from others.